requirements.txt
langchain
langchain-openai
langchain-groq
langchain-community
python-dotenv
streamlit
pydantic
langsmith
python-dotenv
streamlit
faiss-cpu
tiktoken